+++ p2panda-rs/src/entry/tests.rs
@@ -10,7 +10,7 @@ use rstest_reuse::apply;
 use crate::entry::{decode_entry, sign_and_encode, Entry, LogId, SeqNum};
 use crate::identity::KeyPair;
 use crate::operation::{AsOperation, Operation, OperationEncoded};
-use crate::test_utils::fixtures::templates::{many_valid_entries, version_fixtures};
+use crate::test_utils::templates::{many_valid_entries, version_fixtures};
 use crate::test_utils::fixtures::{key_pair, Fixture};
 
 /// Test encoding and decoding entries.
